bottomSheet added
bottom sheet added (the on pressed action (screen)) floating screen
and refactored the project to handle the bottom sheet
