--- lib/widgets/user_transaction.dart
@@ -1,50 +1,0 @@
-import 'package:flutter/material.dart';
-import 'new_transaction.dart';
-import 'transaction_list.dart';
-import '../models/transaction.dart';
-
-class UserTransaction extends StatefulWidget {
-  const UserTransaction({Key? key}) : super(key: key);
-
-  @override
-  State<UserTransaction> createState() => _UserTransactionState();
-}
-
-class _UserTransactionState extends State<UserTransaction> {
-  final List<Transaction> _userTransactions = [
-    Transaction(
-      id: '1',
-      title: 'Big Tasty',
-      amount: 75,
-      date: DateTime.now(),
-    ),
-    Transaction(
-      id: '2',
-      title: '50 Trident Gum',
-      amount: 25,
-      date: DateTime.now(),
-    )
-  ];
-
-  void _addNewTrans(String txTitle, double txAmount) {
-    final newTx = Transaction(
-      id: DateTime.now().toString(),
-      title: txTitle,
-      amount: txAmount,
-      date: DateTime.now(),
-    );
-    setState(() {
-      _userTransactions.add(newTx);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: <Widget>[
-        NewTransaction(_addNewTrans),
-        TransactionList(_userTransactions),
-      ],
-    );
-  }
-}
